fixed spelling error, removed unused imports and moved some initial util methods to util.rs file
