+++ src/orderbook.rs
@@ -1,11 +1,8 @@
 use near_sdk::{
-	env,
 	collections::{
 		UnorderedMap,
 		TreeMap,
-		Vector
 	},
-	json_types::{U128, U64},
 	borsh::{
 		self, 
 		BorshDeserialize, 
@@ -14,10 +11,7 @@ use near_sdk::{
 };
 use std::{
 	cmp,
-	convert::TryInto,
-	collections::HashMap
 };
-use serde_json::json;
 
 /* Import order impl */
 use crate::order;
